1263 - Remove old veteran status card pdf and the feature flag toggling between the old and new
