+++ src/platform/pdf/templates/index.js
@@ -16,10 +16,6 @@ templates.veteranStatus = () => {
   return require('./veteran_status');
 };
 
-templates.veteranStatusNew = () => {
-  return require('./veteran_status_new');
-};
-
 templates.selfEnteredInfo = () => {
   return require('./self_entered_info');
 };
